requirements.txt
easydict==1.9
matplotlib>=3.2.2
networkx>=2.4
numpy>=1.21.6
ogb>=1.3.3
pandas>=1.3.5
PyYAML==3.13
tensorboardX>=2.0
tqdm>=4.64.0
